requirements.txt
streamlit
google-genai
python-dotenv
pyrebase4